refactor: api controller와 view controller를 분리
